refactor: move package - move dto, enum, id to model
