+++ src/main/java/com/group2/glamping/auth/AuthenticationService.java
@@ -4,10 +4,10 @@ package com.group2.glamping.auth;
 import com.group2.glamping.exception.AppException;
 import com.group2.glamping.exception.ErrorCode;
 import com.group2.glamping.model.entity.User;
-import com.group2.glamping.model.entity.dto.requests.AuthenticationRequest;
-import com.group2.glamping.model.entity.dto.requests.RegisterRequest;
-import com.group2.glamping.model.entity.dto.response.AuthenticationResponse;
-import com.group2.glamping.model.entity.enums.Role;
+import com.group2.glamping.model.dto.requests.AuthenticationRequest;
+import com.group2.glamping.model.dto.requests.RegisterRequest;
+import com.group2.glamping.model.dto.response.AuthenticationResponse;
+import com.group2.glamping.model.enums.Role;
 import com.group2.glamping.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
