+++ src/main/java/com/group2/glamping/model/dto/requests/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.group2.glamping.model.entity.dto.requests;
+package com.group2.glamping.model.dto.requests;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
