+++ src/main/java/com/group2/glamping/model/entity/BookingDetailService.java
@@ -1,6 +1,6 @@
 package com.group2.glamping.model.entity;
 
-import com.group2.glamping.model.entity.id.IdBookingDetailService;
+import com.group2.glamping.model.id.IdBookingDetailService;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
