+++ src/main/java/com/group2/glamping/model/entity/User.java
@@ -1,6 +1,6 @@
 package com.group2.glamping.model.entity;
 
-import com.group2.glamping.model.entity.enums.Role;
+import com.group2.glamping.model.enums.Role;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
