+++ src/main/java/com/group2/glamping/model/enums/Role.java
@@ -1,4 +1,4 @@
-package com.group2.glamping.model.entity.enums;
+package com.group2.glamping.model.enums;
 
 public enum Role {
     ROLE_USER, ROLE_ADMIN, ROLE_MANAGER, ROLE_STAFF;
